Shift toggle connection inside src/sso
